+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.0] - November 5, 2018
+
+* **breaking:** once canceled the observable/debouncer is no longer usable.
+* Moved `cancel()` to base class.
+* Return type of callback function is now `void` instead of `dynamic`.
+* Add more to `./example/main.dart`.
+
 ## [0.1.1] - October 22, 2018
 
 * Add `setValue()`.